pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serena-cli"
version = "1.0.1"
description = "A powerful CLI tool for quickly enabling and configuring Serena coding agent tools in specified projects"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Panda", email = "panda@example.com"}
]
maintainers = [
    {name = "Panda", email = "panda@example.com"}
]
keywords = ["serena", "cli", "mcp", "coding-agent", "ai-tools", "project-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Utilities",
    "Environment :: Console",
]
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
]

[project.scripts]
serena-cli = "serena_cli.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/serena-cli"
Documentation = "https://github.com/yourusername/serena-cli#readme"
Repository = "https://github.com/yourusername/serena-cli"
Issues = "https://github.com/yourusername/serena-cli/issues"
Changelog = "https://github.com/yourusername/serena-cli/blob/main/CHANGELOG.md"
Download = "https://pypi.org/project/serena-cli/#files"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=serena_cli --cov-report=term-missing"
